Update to retrieve database information from AWS SSM
Instead of expecting the information needed to build a MongoDB URI to
be provided through environment variables we instead expect AWS SSM
Parameter Store keys to be passed and the values of those keys to
provide the required information. This is in line with other AWS
Lambdas we have and improves maintainability.